// src/models/routes.ts
export enum appRoutePath {
    home = "/home",
    availability = "/availability",
    promotions = "/promotions",
    reservations = "/reservations",
    property = "/property",
    opportunities = "/opportunities",
    inbox = "/inbox",
    reviews = "/reviews",
    finance = "/finance",
    analytics = "/analytics"
}